Add files via upload
Analysis 1: I tried applying my limited skill-set of Data Analysis to a Covid-19 Data-set of India, I downloaded from Kaggle. The data-set has information recorded about every patient of covid-19 including details like their state, the date on which they were diagnosed with the virus, their current status (recovered/hospitalised), etc. The purpose of this exercise was to check if my results were even remotely close to those that were uploaded by the Covid-19 tracker of India. I have also tried showing the percentage increase in the number of cases per day and when it initially hit its peak. 

Analysis 2: In the initial stages of the virus transmission, it was believed that males are more likely to be affected than females. I tried testing the hypothesis, trying to figure out if that was really the case. I calculated the results for 51 countries and in majority of the countries, males were more adversely affected than females. 

Analysis 3: Scientists all around the world have been trying to test a hypothesis that states ‘Transmission of Covid-19 infections tend to slow down in high heat.’ So, as a part of a learning exercise and to apply my rudimentary knowledge about Data Analysis, I collected data of the top 6 hottest states of the country and formulated a basic code that would give me the ‘Average increase in the number of Covid-19 cases per day’ in the month of May. Later on, I compared the obtained quantities to the ‘Average temperature of the State in the month of May’, respectively.

The result below shows that ‘Rajasthan’ recorded the highest temperature. So as per our premise, it should have the least value for ‘Average Increase in the number of cases per day’. Although, it is quite not the case.